refactor: Move each message box to its own file
